// frontend/src/api/generated/model/catalogAssetListSort.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Geo Portal API
 * Internal API for the Geo Portal SPA. Read-only over the knowledge base.
 * OpenAPI spec version: 0.1.0
 */

export type CatalogAssetListSort = typeof CatalogAssetListSort[keyof typeof CatalogAssetListSort];


export const CatalogAssetListSort = {
  '-ingested_at': '-ingested_at',
  '-name': '-name',
  data_type: 'data_type',
  ingested_at: 'ingested_at',
  name: 'name',
} as const;